requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
